backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
python-dotenv
requests
Pillow
whitenoise
embit>=0.8.0